core/auth.py
from typing import Union, Any
from datetime import datetime, timedelta

from jose import jwt
from passlib.context import CryptContext

from settings import settings


AUTH = settings.AUTH
password_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
# TODO: 密码加盐


def get_hashed_password(password: str) -> str:
    return password_context.hash(password)


def verify_password(password: str, hashed_pass: str) -> bool:
    return password_context.verify(password, hashed_pass)


def create_access_token(subject: Union[str, Any], expires_delta: int = None) -> str:
    if expires_delta is not None:
        expires_delta = datetime.utcnow() + expires_delta
    else:
        expires_delta = datetime.utcnow() + timedelta(minutes=AUTH.ACCESS_TOKEN_EXPIRE_MINUTES)

    to_encode = {"exp": expires_delta, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, AUTH.JWT_SECRET_KEY, AUTH.ALGORITHM)
    return encoded_jwt


def create_refresh_token(subject: Union[str, Any], expires_delta: int = None) -> str:
    if expires_delta is not None:
        expires_delta = datetime.utcnow() + expires_delta
    else:
        expires_delta = datetime.utcnow() + timedelta(minutes=AUTH.REFRESH_TOKEN_EXPIRE_MINUTES)

    to_encode = {"exp": expires_delta, "sub": str(subject)}
    encoded_jwt = jwt.encode(
        to_encode, AUTH.JWT_REFRESH_SECRET_KEY, AUTH.ALGORITHM)
    return encoded_jwt
